--- src/components/UserRecent.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>Soon...</div>
-</template>
-
-<script lang="ts">
-import Vue from 'vue';
-import { Component } from 'vue-property-decorator';
-
-@Component
-export default class UserRecent extends Vue {}
-</script>
