+++ src/router/index.ts
@@ -23,19 +23,9 @@ const routes: Array<RouteConfig> = [
     }
   },
   {
-    path: '/user',
+    path: '/user/:id',
     name: 'User',
     component: () => import(/* webpackChunkName: "retrive" */ '../views/TheUser.vue'),
-    children: [
-      {
-        path: '/user',
-        component: () =>import(/* webpackChunkName: "Overview" */ '../components/UserOverview.vue'),
-      },
-      {
-        path: '/user/userrecent',
-        component: () =>import(/* webpackChunkName: "Overview" */ '../components/UserRecent.vue'),
-      },
-    ],
   },
   {
     path: "/tournament/:id/:module?",
